requirements.txt
anthropic>=0.40.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
streamlit>=1.40.0
python-dotenv>=1.0.0